[storybook Rename all __test__ references to __tests__
