+++ src/Apps/Artwork/Components/__stories__/ArtistInfo.story.tsx
@@ -1,7 +1,7 @@
 import { Box } from "@artsy/palette"
-// import { bio, biography_blurb } from "Apps/__test__/Fixtures/ArtistBio"
-// import { artistResponse } from "Apps/__test__/Fixtures/MarketInsights"
-// import { exhibitions } from "Apps/__test__/Fixtures/SelectedExhibitions"
+// import { bio, biography_blurb } from "Apps/__tests__/Fixtures/ArtistBio"
+// import { artistResponse } from "Apps/__tests__/Fixtures/MarketInsights"
+// import { exhibitions } from "Apps/__tests__/Fixtures/SelectedExhibitions"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
 import { Section } from "Styleguide/Utils/Section"
