+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebar/ArtworkSidebarExtraLinks.story.tsx
@@ -6,7 +6,7 @@ import {
   ForSaleArtworkWithMultipleConsignableArtists,
   ForSaleArtworkWithNoConsignableArtists,
   ForSaleArtworkWithOneConsignableArtist,
-} from "Apps/__test__/Fixtures/Artwork/ArtworkSidebar/ArtworkSidebarExtraLinks"
+} from "Apps/__tests__/Fixtures/Artwork/ArtworkSidebar/ArtworkSidebarExtraLinks"
 import { ArtworkSidebarExtraLinks as ExtraLinks } from "Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarExtraLinks"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
