+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebar/ArtworkSidebarMetadata.story.tsx
@@ -6,7 +6,7 @@ import {
   FilledOutMetadataNoEditions,
   FilledOutMetadataOneEditionSet,
   MetadataForAuctionWork,
-} from "Apps/__test__/Fixtures/Artwork/ArtworkSidebar/ArtworkSidebarMetadata"
+} from "Apps/__tests__/Fixtures/Artwork/ArtworkSidebar/ArtworkSidebarMetadata"
 import { ArtworkSidebarMetadata as Metadata } from "Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarMetadata"
 import { RelayStubProvider } from "DevTools/RelayStubProvider"
 import React from "react"
