+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebar/ArtworkSidebarSizeInfo.story.tsx
@@ -2,7 +2,7 @@ import {
   ArtworkWithEditionOfOnly,
   ArtworkWithSizeAndEditionOf,
   ArtworkWithSizeOnly,
-} from "Apps/__test__/Fixtures/Artwork/ArtworkSidebar/ArtworkSidebarSizeInfo"
+} from "Apps/__tests__/Fixtures/Artwork/ArtworkSidebar/ArtworkSidebarSizeInfo"
 import { ArtworkSidebarSizeInfo as SizeInfo } from "Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarSizeInfo"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
