+++ src/Apps/Artwork/Components/__stories__/ArtworkSidebar/ArtworkSidebarTitleInfo.story.tsx
@@ -3,7 +3,7 @@ import {
   ArtworkWithTitleAndMedium,
   ArtworkWithTitleDateAndMedium,
   ArtworkWithTitleOnly,
-} from "Apps/__test__/Fixtures/Artwork/ArtworkSidebar/ArtworkSidebarTitleInfo"
+} from "Apps/__tests__/Fixtures/Artwork/ArtworkSidebar/ArtworkSidebarTitleInfo"
 import { ArtworkSidebarTitleInfo as TitleInfo } from "Apps/Artwork/Components/ArtworkSidebar/ArtworkSidebarTitleInfo"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
