+++ src/Styleguide/Components/__stories__/Carousel.story.tsx
@@ -1,5 +1,5 @@
 import { Box, Image } from "@artsy/palette"
-import { artworkBricks, images } from "Apps/__test__/Fixtures/Carousel"
+import { artworkBricks, images } from "Apps/__tests__/Fixtures/Carousel"
 import { FillwidthItem } from "Components/Artwork/FillwidthItem"
 import React from "react"
 import { storiesOf } from "storybook/storiesOf"
